+++ src/main/java/com/spartaschedulerdevelop/filter/LoginFilter.java
@@ -44,7 +44,7 @@ public class LoginFilter implements Filter {
             HttpSession session = httpRequest.getSession(false);
 
             // 로그인하지 않은 사용자인 경우
-            if (session == null || session.getAttribute("sessionKey값") == null) {
+            if (session == null || session.getAttribute("userId") == null) {
                 ErrorResponse errorResponse = ErrorResponse.of(
                         HttpStatus.UNAUTHORIZED,
                         "AUTH-004",
